mooore refactoring
