+++ Projects/EditCoInsured/Sources/View/Components/ConfirmChangesView.swift
@@ -1,0 +1,43 @@
+import SwiftUI
+import hCore
+import hCoreUI
+
+struct ConfirmChangesView: View {
+    @ObservedObject private var editCoInsuredNavigation: EditCoInsuredNavigationViewModel
+    @ObservedObject var intentViewModel: IntentViewModel
+
+    public init(
+        editCoInsuredNavigation: EditCoInsuredNavigationViewModel
+    ) {
+        self.editCoInsuredNavigation = editCoInsuredNavigation
+        self.intentViewModel = editCoInsuredNavigation.intentViewModel
+    }
+
+    var body: some View {
+        hSection {
+            VStack(spacing: .padding16) {
+                PriceField(
+                    newPremium: intentViewModel.intent.newPremium,
+                    currentPremium: intentViewModel.intent.currentPremium,
+                    subTitle: L10n.contractAddCoinsuredStartsFrom(
+                        intentViewModel.intent.activationDate.localDateToDate?.displayDateDDMMMYYYYFormat ?? ""
+                    )
+                )
+                .hWithStrikeThroughPrice(setTo: .crossOldPrice)
+
+                hButton.LargeButton(type: .primary) {
+                    editCoInsuredNavigation.showProgressScreenWithSuccess = true
+                    Task {
+                        await intentViewModel.performCoInsuredChanges(
+                            commitId: intentViewModel.intent.id
+                        )
+                    }
+                } content: {
+                    hText(L10n.contractAddCoinsuredConfirmChanges)
+                }
+                .hButtonIsLoading(intentViewModel.isLoading)
+            }
+        }
+        .sectionContainerStyle(.transparent)
+    }
+}
